+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 - Add support for passing config params `maxEdge` and `maxTextSize` to mermaid by [@ad1992](https://github.com/ad1992) in https://github.com/excalidraw/mermaid-to-excalidraw/pull/68
 
-Additonally the param `fontSize` is renamed to `themeVariables.fontSize` to be consistent with the [mermaid config](https://mermaid.js.org/schemas/config.schema.json).
+Additonally the param `fontSize` is renamed to `themeVariables.fontSize` and type is changed from `number` to `string` to be consistent with the [mermaid config](https://mermaid.js.org/schemas/config.schema.json).
 
 ## v1.0.0 (2024-05-20)